Added sample caller and callee projects
The caller's executable will be loaded in the Sim Harness GUI. But
first, the path to the callee is hardcoded in the caller's main
function, so it needs to be changed and then rebuilt. The caller will
automatically start the callee and redirect its stdin, stdout, and
stderr to streams defined in the caller. Once the setup is complete, the
caller goes into an infinite loop listening for the stdout of the
callee.